+++ src/UiCalendarMonth.vue
@@ -1,20 +1,18 @@
 <template>
-    <div class="ui-calendar-month">
-        <div class="ui-calendar-month__header">
-            <span v-for="day in daysOfWeek">{{ day }}</span>
-        </div>
+    <table class="ui-calendar-month">
+        <thead class="ui-calendar-month__header">
+            <tr>
+                <th v-for="day in daysOfWeek">{{ day }}</th>
+            </tr>
+        </thead>
 
-        <div
-            class="ui-calendar-month__week is-current"
-            ref="current"
+        <tbody class="ui-calendar-month__body">
+            <tr
+                is="ui-calendar-week"
 
-            :class="weekClasses"
-
-            @transitionend="onTransitionEnd"
-        >
-            <ui-calendar-week
+                :color="color"
                 :date-filter="dateFilter"
-                :key="index"
+                :key="date.toString()"
                 :max-date="maxDate"
                 :min-date="minDate"
                 :month="currentWeekStartDates[1].getMonth()"
@@ -23,26 +21,10 @@
 
                 @date-select="onDateSelect"
 
-                v-for="(date, index) in currentWeekStartDates"
-            ></ui-calendar-week>
-        </div>
-
-        <div class="ui-calendar-month__week is-other" ref="other" :class="weekClasses">
-            <ui-calendar-week
-                :key="index"
-                :max-date="maxDate"
-                :min-date="minDate"
-                :month="otherWeekStartDates[1].getMonth()"
-                :selected="selected"
-                :visible="false"
-                :week-start="date"
-
-                @date-select="onDateSelect"
-
-                v-for="(date, index) in otherWeekStartDates"
-            ></ui-calendar-week>
-        </div>
-    </div>
+                v-for="date in currentWeekStartDates"
+            ></tr>
+        </tbody>
+    </table>
 </template>
 
 <script>
@@ -63,19 +45,11 @@ export default {
         startOfWeek: {
             type: Number,
             default: 0
+        },
+        color: {
+            type: String,
+            default: 'primary' // 'primary' or 'accent'
         }
-    },
-
-    data() {
-        return {
-            dateOutOfView: dateUtils.clone(this.dateInView),
-            isSliding: false,
-            slideDirection: '',
-
-            // Detects IE and not Edge: http://stackoverflow.com/a/22082397
-            isIE: Boolean(window.MSInputMethodContext) && Boolean(document.documentMode),
-            ieTimeout: null
-        };
     },
 
     computed: {
@@ -91,19 +65,8 @@ export default {
             return days.concat(this.lang.days.initials.slice(0, this.startOfWeek));
         },
 
-        weekClasses() {
-            return [
-                { [`ui-calendar-month--slide-${this.slideDirection}`]: this.isSliding },
-                { 'is-sliding': this.isSliding }
-            ];
-        },
-
         currentWeekStartDates() {
             return this.getWeekStartDates(this.dateInView);
-        },
-
-        otherWeekStartDates() {
-            return this.getWeekStartDates(this.dateOutOfView);
         }
     },
 
@@ -128,39 +91,12 @@ export default {
             return starts;
         },
 
-        goToDate(date, options = { isForward: true }) {
-            this.isSliding = true;
-            this.slideDirection = options.isForward ? 'left' : 'right';
-            this.dateOutOfView = dateUtils.clone(date);
-
-            // A hack for IE: sometimes when rapidly scrolling through months, the
-            // transitionend event is not fired, causing the month to not change.
-            // This ensures that onTransitionEnd() is called after 300ms.
-            if (this.isIE) {
-                this.ieTimeout = setTimeout(this.onTransitionEnd, 300);
-            }
+        goToDate(date) {
+            this.$emit('change', dateUtils.clone(date));
         },
 
         onDateSelect(date) {
             this.$emit('date-select', date);
-        },
-
-        onTransitionEnd() {
-            if (this.ieTimeout) {
-                clearTimeout(this.ieTimeout);
-                this.ieTimeout = null;
-
-                // Abort if the transition has already ended
-                if (!this.isSliding) {
-                    return;
-                }
-            }
-
-            this.isSliding = false;
-            this.slideDirection = '';
-
-            this.$emit('change', dateUtils.clone(this.dateOutOfView));
-            this.$emit('transition-end');
         }
     },
 
@@ -174,58 +110,26 @@ export default {
 @import './styles/imports';
 
 .ui-calendar-month {
-    height: ($ui-calendar-cell-size * 6) + $ui-calendar-month-header-height;
-    overflow: hidden;
-    position: relative;
+    table-layout: fixed;
     width: 100%;
 }
 
 .ui-calendar-month__header {
-    align-items: center;
-    display: flex;
-    height: $ui-calendar-month-header-height;
-    justify-content: space-around;
     width: 100%;
 
-    span {
-        align-items: center;
+    th {
         color: $secondary-text-color;
-        display: flex;
         font-size: rem(14px);
-        height: $ui-calendar-cell-size;
-        justify-content: center;
+        font-weight: 600;
+        height: $ui-calendar-month-header-height;
+        text-align: center;
         text-transform: uppercase;
-        width: $ui-calendar-cell-size;
+        vertical-align: middle;
+        width: (100% / 7);
     }
 }
 
-.ui-calendar-month__week {
-    $week-width: $ui-calendar-cell-size * 7;
-    position: absolute;
-    width: $week-width;
-
-    &.is-current {
-        left: 0;
-    }
-
-    &.is-other {
-        left: $week-width;
-
-        &.ui-calendar-month--slide-right {
-            left: -$week-width;
-        }
-    }
-
-    &.is-sliding {
-        transition: transform 250ms ease;
-
-        &.ui-calendar-month--slide-left {
-            transform: translate3d(-$week-width, 0, 0);
-        }
-
-        &.ui-calendar-month--slide-right {
-            transform: translate3d($week-width, 0, 0);
-        }
-    }
+.ui-calendar-month__body {
+    width: 100%;
 }
-</style>s
+</style>